Performance testing squared distance vs distance

+++ tsp_utils.c
@@ -3,8 +3,8 @@
 
 int calculate_squared_distance(const City *city1, const City *city2)
 {
-    int dx = city1->x - city2->x;
-    int dy = city1->y - city2->y;
+    double dx = city1->x - city2->x;
+    double dy = city1->y - city2->y;
     return dx*dx + dy*dy;
 }
